Rename C.create -> C.save_wasm and test
